requirements.txt
asgiref==3.7.2
Django==4.2.4
mysql-connector==2.2.9
mysql-connector-python==8.1.0
protobuf==4.21.12
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.7.1
